safari-600.7-branch/LayoutTests/platform/ios-sim/css3/device-adapt/opera/constrain-006-expected.txt


FAIL CSS Test: @viewport constrained - width 200% assert_equals: expected 1600 but got 800(stack: assert@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:2168:73
assert_equals@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:705:15
file:///Volumes/Data/WebKit.git/LayoutTests/css3/device-adapt/opera/constrain-006.xhtml:69:22
step@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:1186:30
onload@file:///Volumes/Data/WebKit.git/LayoutTests/css3/device-adapt/opera/constrain-006.xhtml:68:14)
